refactor: remove database.ts duplication, consolidate to entities.ts
Deleted: lib/types/database.ts (78 lines of duplicate type definitions)

Updated imports across codebase:
- app/api/posts/route.ts: CreatePost → CreatePostDTO
- app/api/posts/[id]/route.ts: UpdatePost → UpdatePostDTO
- app/api/auth/signup/route.ts: SignUpRequest from entities
- app/api/auth/signin/route.ts: SignInRequest from entities
- components/create-post-form.tsx: types from entities
- components/feed.tsx: PostWithProfile from entities
- components/post-card.tsx: PostWithProfile from entities

Before:
- database.ts: Manual type definitions (duplication)
- supabase.ts: Auto-generated types (duplication)
- entities.ts: More manual definitions (triplication!)

After:
- supabase.ts: Auto-generated (SINGLE SOURCE OF TRUTH)
- entities.ts: Extracts from supabase.ts + DTOs

Result:
✅ Single source of truth for DB types
✅ Deleted 78 lines of duplicate code
✅ All types auto-sync with schema
