+++ app/api/auth/signup/route.ts
@@ -1,6 +1,6 @@
 import { createClient } from "@/lib/supabase/server";
 import { NextRequest, NextResponse } from "next/server";
-import { SignUpRequest } from "@/lib/types/database";
+import { SignUpRequest } from "@/lib/types/entities";
 
 /**
  * @swagger
